+++ src/Http/Http/src/Features/ResponseBodyPipeFeature.cs
@@ -12,8 +12,6 @@ namespace Microsoft.AspNetCore.Http.Features
     {
         private PipeWriter _internalPipeWriter;
         private Stream _streamInstanceWhenWrapped;
-
-        private PipeWriter _userSetPipeWriter;
         private HttpContext _context;
 
         public ResponseBodyPipeFeature(HttpContext context)
@@ -29,37 +27,21 @@ namespace Microsoft.AspNetCore.Http.Features
         {
             get
             {
-                if (_userSetPipeWriter != null)
-                {
-                    return _userSetPipeWriter;
-                }
-
                 if (_internalPipeWriter == null ||
                     !ReferenceEquals(_streamInstanceWhenWrapped, _context.Response.Body))
                 {
                     _streamInstanceWhenWrapped = _context.Response.Body;
                     _internalPipeWriter = PipeWriter.Create(_context.Response.Body);
 
-                    _context.Response.OnCompleted((self) => ((ResponseBodyPipeFeature)self).Complete(), this);
+                    _context.Response.OnCompleted((self) =>
+                    {
+                        ((PipeWriter)self).Complete();
+                        return Task.CompletedTask;
+                    }, _internalPipeWriter);
                 }
 
                 return _internalPipeWriter;
             }
-            set
-            {
-                _userSetPipeWriter = value ?? throw new ArgumentNullException(nameof(value));
-                // REVIEW: Should we do this (vvvv) or just get rid of the setter?
-                // TODO set the response body Stream to an adapted pipe https://github.com/aspnet/AspNetCore/issues/3971
-            }
-        }
-
-        private Task Complete()
-        {
-            if(_internalPipeWriter != null && ReferenceEquals(_context.Response.Body, _streamInstanceWhenWrapped))
-            {
-                _internalPipeWriter.Complete();
-            }
-            return Task.CompletedTask;
         }
     }
 }
