--- src/Http/Http/test/FlushResultCancellationTests.cs
@@ -1,30 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using System.IO.Pipelines;
-using System.Runtime.CompilerServices;
-using System.Threading;
-using System.Threading.Tasks;
-using Xunit;
-
-namespace System.IO.Pipelines.Tests
-{
-    public class FlushResultCancellationTests : StreamPipeTest
-    {
-        [Fact]
-        public async Task FlushAsyncWithNewCancellationTokenNotAffectedByPrevious()
-        {
-            var cancellationTokenSource1 = new CancellationTokenSource();
-            PipeWriter buffer = Writer.WriteEmpty(10);
-            await buffer.FlushAsync(cancellationTokenSource1.Token);
-
-            cancellationTokenSource1.Cancel();
-
-            var cancellationTokenSource2 = new CancellationTokenSource();
-            buffer = Writer.WriteEmpty(10);
-
-            await buffer.FlushAsync(cancellationTokenSource2.Token);
-        }
-    }
-}
