+++ src/Servers/Kestrel/Core/src/Internal/Http/HttpResponseStream.cs
@@ -2,33 +2,54 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.IO;
 using System.IO.Pipelines;
+using System.Threading;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http.Features;
+using Microsoft.AspNetCore.Internal;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
 {
-    internal sealed class HttpResponseStream : WriteOnlyPipeStream
+    internal sealed class HttpResponseStream : Stream
     {
         private readonly HttpResponsePipeWriter _pipeWriter;
         private readonly IHttpBodyControlFeature _bodyControl;
 
         public HttpResponseStream(IHttpBodyControlFeature bodyControl, HttpResponsePipeWriter pipeWriter)
-            : base(pipeWriter)
         {
             _bodyControl = bodyControl;
             _pipeWriter = pipeWriter;
         }
 
-        public override void Write(byte[] buffer, int offset, int count)
-        {
-            if (!_bodyControl.AllowSynchronousIO)
-            {
-                throw new InvalidOperationException(CoreStrings.SynchronousWritesDisallowed);
-            }
+        public override bool CanSeek => false;
 
-            base.Write(buffer, offset, count);
+        public override bool CanRead => false;
+        
+        public override bool CanWrite => true;
+        
+        public override long Length => throw new NotSupportedException();
+
+        public override long Position
+        {
+            get => throw new NotSupportedException();
+            set => throw new NotSupportedException();
+        }
+        
+        public override int ReadTimeout
+        {
+            get => throw new NotSupportedException();
+            set => throw new NotSupportedException();
         }
 
+        public PipeWriter InnerPipeWriter { get; }
+
+        public override int Read(byte[] buffer, int offset, int count)
+            => throw new NotSupportedException();
+
+        public override Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
+          => throw new NotSupportedException();
+        
         public override void Flush()
         {
             if (!_bodyControl.AllowSynchronousIO)
@@ -36,7 +57,87 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
                 throw new InvalidOperationException(CoreStrings.SynchronousWritesDisallowed);
             }
 
-            base.Flush();
+            FlushAsync(default).GetAwaiter().GetResult();
+        }
+        
+        public override Task FlushAsync(CancellationToken cancellationToken)
+        {
+            return InnerPipeWriter.FlushAsync(cancellationToken).GetAsTask();
+        }
+        
+        public override long Seek(long offset, SeekOrigin origin)
+        {
+            throw new NotSupportedException();
+        }
+
+        
+        public override void SetLength(long value)
+        {
+            throw new NotSupportedException();
+        }
+
+        
+        public override void Write(byte[] buffer, int offset, int count)
+        {
+            if (!_bodyControl.AllowSynchronousIO)
+            {
+                throw new InvalidOperationException(CoreStrings.SynchronousWritesDisallowed);
+            }
+
+            WriteAsync(buffer, offset, count, default).GetAwaiter().GetResult();
+        }
+
+        public override IAsyncResult BeginWrite(byte[] buffer, int offset, int count, AsyncCallback callback, object state)
+        {
+            var task = WriteAsync(buffer, offset, count, default, state);
+            if (callback != null)
+            {
+                task.ContinueWith(t => callback.Invoke(t));
+            }
+            return task;
+        }
+        
+        public override void EndWrite(IAsyncResult asyncResult)
+        {
+            ((Task<object>)asyncResult).GetAwaiter().GetResult();
+        }
+
+        private Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken, object state)
+        {
+            var tcs = new TaskCompletionSource<object>(state);
+            var task = WriteAsync(buffer, offset, count, cancellationToken);
+            task.ContinueWith((task2, state2) =>
+            {
+                var tcs2 = (TaskCompletionSource<object>)state2;
+                if (task2.IsCanceled)
+                {
+                    tcs2.SetCanceled();
+                }
+                else if (task2.IsFaulted)
+                {
+                    tcs2.SetException(task2.Exception);
+                }
+                else
+                {
+                    tcs2.SetResult(null);
+                }
+            }, tcs, cancellationToken);
+            return tcs.Task;
+        }
+
+        public override Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
+        {
+            return WriteAsyncInternal(new ReadOnlyMemory<byte>(buffer, offset, count), cancellationToken);
+        }
+        
+        public override ValueTask WriteAsync(ReadOnlyMemory<byte> source, CancellationToken cancellationToken = default)
+        {
+            return new ValueTask(WriteAsyncInternal(source, cancellationToken));
+        }
+
+        private Task WriteAsyncInternal(ReadOnlyMemory<byte> source, CancellationToken cancellationToken = default)
+        {
+            return InnerPipeWriter.WriteAsync(source, cancellationToken).GetAsTask();
         }
     }
 }
