+++ src/Servers/Kestrel/test/InMemory.FunctionalTests/RequestTests.cs
@@ -58,30 +58,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.InMemory.FunctionalTests
         }
 
         [Fact]
-        public async Task PipesAreNotPersistedBySettingStreamPipeWriterAcrossRequests()
-        {
-            var responseBodyPersisted = false;
-            PipeWriter bodyPipe = null;
-            await using (var server = new TestServer(async context =>
-            {
-                if (context.Response.BodyWriter == bodyPipe)
-                {
-                    responseBodyPersisted = true;
-                }
-                bodyPipe = new StreamPipeWriter(new MemoryStream());
-                context.Response.BodyWriter = bodyPipe;
-
-                await context.Response.WriteAsync("hello, world");
-            }, new TestServiceContext(LoggerFactory)))
-            {
-                Assert.Equal(string.Empty, await server.HttpClientSlim.GetStringAsync($"http://localhost:{server.Port}/"));
-                Assert.Equal(string.Empty, await server.HttpClientSlim.GetStringAsync($"http://localhost:{server.Port}/"));
-
-                Assert.False(responseBodyPersisted);
-            }
-        }
-
-        [Fact]
         public async Task PipesAreNotPersistedAcrossRequests()
         {
             var responseBodyPersisted = false;
