+++ src/Servers/Kestrel/test/InMemory.FunctionalTests/ResponseTests.cs
@@ -3852,35 +3852,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.InMemory.FunctionalTests
         }
 
         [Fact]
-        public async Task ResponseSetBodyAndPipeBodyIsWrapped()
-        {
-            await using (var server = new TestServer(async httpContext =>
-            {
-                httpContext.Response.Body = new MemoryStream();
-                httpContext.Response.BodyWriter = new Pipe().Writer;
-                Assert.IsType<WriteOnlyPipeStream>(httpContext.Response.Body);
-
-                await Task.CompletedTask;
-            }, new TestServiceContext(LoggerFactory)))
-            {
-                using (var connection = server.CreateConnection())
-                {
-                    await connection.Send(
-                        "GET / HTTP/1.1",
-                        "Host:",
-                        "",
-                        "");
-                    await connection.Receive(
-                        "HTTP/1.1 200 OK",
-                        $"Date: {server.Context.DateHeaderValue}",
-                        "Content-Length: 0",
-                        "",
-                        "");
-                }
-            }
-        }
-
-        [Fact]
         public async Task ResponseSetBodyToSameValueTwiceGetPipeMultipleTimesDifferentObject()
         {
             await using (var server = new TestServer(async httpContext =>
@@ -3894,181 +3865,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.InMemory.FunctionalTests
 
                 Assert.NotEqual(BodyWriter1, BodyWriter2);
                 await Task.CompletedTask;
-            }, new TestServiceContext(LoggerFactory)))
-            {
-                using (var connection = server.CreateConnection())
-                {
-                    await connection.Send(
-                        "GET / HTTP/1.1",
-                        "Host:",
-                        "",
-                        "");
-                    await connection.Receive(
-                        "HTTP/1.1 200 OK",
-                        $"Date: {server.Context.DateHeaderValue}",
-                        "Content-Length: 0",
-                        "",
-                        "");
-                }
-            }
-        }
-
-        [Fact]
-        public async Task ResponseSetPipeToSameValueTwiceGetBodyMultipleTimesDifferent()
-        {
-            await using (var server = new TestServer(async httpContext =>
-            {
-                var pipeWriter = new Pipe().Writer;
-                httpContext.Response.BodyWriter = pipeWriter;
-                var body1 = httpContext.Response.Body;
-
-                httpContext.Response.BodyWriter = pipeWriter;
-                var body2 = httpContext.Response.Body;
-
-                Assert.NotEqual(body1, body2);
-                await Task.CompletedTask;
-            }, new TestServiceContext(LoggerFactory)))
-            {
-                using (var connection = server.CreateConnection())
-                {
-                    await connection.Send(
-                        "GET / HTTP/1.1",
-                        "Host:",
-                        "",
-                        "");
-                    await connection.Receive(
-                        "HTTP/1.1 200 OK",
-                        $"Date: {server.Context.DateHeaderValue}",
-                        "Content-Length: 0",
-                        "",
-                        "");
-                }
-            }
-        }
-
-        [Fact]
-        public async Task ResponseSetPipeAndBodyWriterIsWrapped()
-        {
-            await using (var server = new TestServer(async httpContext =>
-            {
-                httpContext.Response.BodyWriter = new Pipe().Writer;
-                httpContext.Response.Body = new MemoryStream();
-                Assert.IsType<StreamPipeWriter>(httpContext.Response.BodyWriter);
-                Assert.IsType<MemoryStream>(httpContext.Response.Body);
-
-                await Task.CompletedTask;
-            }, new TestServiceContext(LoggerFactory)))
-            {
-                using (var connection = server.CreateConnection())
-                {
-                    await connection.Send(
-                        "GET / HTTP/1.1",
-                        "Host:",
-                        "",
-                        "");
-                    await connection.Receive(
-                        "HTTP/1.1 200 OK",
-                        $"Date: {server.Context.DateHeaderValue}",
-                        "Content-Length: 0",
-                        "",
-                        "");
-                }
-            }
-        }
-
-        [Fact]
-        public async Task ResponseWriteToBodyWriterAndStreamAllBlocksDisposed()
-        {
-            await using (var server = new TestServer(async httpContext =>
-            {
-                for (var i = 0; i < 3; i++)
-                {
-                    httpContext.Response.BodyWriter = new Pipe().Writer;
-                    await httpContext.Response.Body.WriteAsync(new byte[1]);
-                    httpContext.Response.Body = new MemoryStream();
-                    await httpContext.Response.BodyWriter.WriteAsync(new byte[1]);
-                }
-
-                // TestMemoryPool will confirm that all rented blocks have been disposed, meaning dispose was called.
-
-                await Task.CompletedTask;
-            }, new TestServiceContext(LoggerFactory)))
-            {
-                using (var connection = server.CreateConnection())
-                {
-                    await connection.Send(
-                        "GET / HTTP/1.1",
-                        "Host:",
-                        "",
-                        "");
-                    await connection.Receive(
-                        "HTTP/1.1 200 OK",
-                        $"Date: {server.Context.DateHeaderValue}",
-                        "Content-Length: 0",
-                        "",
-                        "");
-                }
-            }
-        }
-
-        [Fact]
-        public async Task ResponseStreamWrappingWorks()
-        {
-            await using (var server = new TestServer(async httpContext =>
-            {
-                var oldBody = httpContext.Response.Body;
-                httpContext.Response.Body = new MemoryStream();
-
-                await httpContext.Response.BodyWriter.WriteAsync(new byte[1]);
-                await httpContext.Response.Body.WriteAsync(new byte[1]);
-
-                Assert.Equal(2, httpContext.Response.Body.Length);
-
-                httpContext.Response.Body = oldBody;
-
-                // Even though we are restoring the original response body, we will create a
-                // wrapper rather than restoring the original pipe.
-                Assert.IsType<StreamPipeWriter>(httpContext.Response.BodyWriter);
-
-            }, new TestServiceContext(LoggerFactory)))
-            {
-                using (var connection = server.CreateConnection())
-                {
-                    await connection.Send(
-                        "GET / HTTP/1.1",
-                        "Host:",
-                        "",
-                        "");
-                    await connection.Receive(
-                        "HTTP/1.1 200 OK",
-                        $"Date: {server.Context.DateHeaderValue}",
-                        "Content-Length: 0",
-                        "",
-                        "");
-                }
-            }
-        }
-
-        [Fact]
-        public async Task ResponsePipeWrappingWorks()
-        {
-            await using (var server = new TestServer(async httpContext =>
-            {
-                var oldPipeWriter = httpContext.Response.BodyWriter;
-                var pipe = new Pipe();
-                httpContext.Response.BodyWriter = pipe.Writer;
-
-                await httpContext.Response.Body.WriteAsync(new byte[1]);
-                await httpContext.Response.BodyWriter.WriteAsync(new byte[1]);
-
-                var readResult = await pipe.Reader.ReadAsync();
-                Assert.Equal(2, readResult.Buffer.Length);
-
-                httpContext.Response.BodyWriter = oldPipeWriter;
-
-                // Even though we are restoring the original response body, we will create a
-                // wrapper rather than restoring the original pipe.
-                Assert.IsType<WriteOnlyPipeStream>(httpContext.Response.Body);
             }, new TestServiceContext(LoggerFactory)))
             {
                 using (var connection = server.CreateConnection())
